recalculate campain

+++ app/campaigns/models.py
@@ -19,7 +19,7 @@ class Campaign(models.Model):
         max_digits=10, decimal_places=2, default=0)
     amount_pending = models.DecimalField(
         max_digits=10, decimal_places=2, default=0)
-    total_ads = models.IntegerField(default=0)
+    total_ads = models.DecimalField(max_digits=13, decimal_places=8, default=0)
     profit = models.DecimalField(max_digits=10, decimal_places=2, default=0)
     ROI = models.DecimalField(max_digits=5, decimal_places=2, default=0)
     total_views = models.IntegerField(default=0)
